Backend_API/code/collection/requirements.txt
boto3
datetime
pandas_datareader
yfinance
pandas-stubs